requirements.txt
fastapi
python-pptx
pydantic
Pillow
uvicorn
python-multipart
